solution

+++ src/main/java/core/basesyntax/strategy/impl/PurchaseOperation.java
@@ -7,6 +7,14 @@ import core.basesyntax.strategy.OperationHandler;
 public class PurchaseOperation implements OperationHandler {
     @Override
     public void apply(FruitTransaction transaction) {
+        int availableQuantity = Storage.storage.get(transaction.getFruit());
+
+        if (availableQuantity < transaction.getQuantity()) {
+            throw new IllegalArgumentException("Not enough "
+                    + transaction.getFruit() + " in stock for purchase. Available: "
+                    + availableQuantity + ", Required: " + transaction.getQuantity());
+        }
+
         Storage.remove(transaction.getFruit(), transaction.getQuantity());
     }
 }
